[UPDATE] implementation
